% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_jdbc.R
\name{databricks_jdbc}
\alias{databricks_jdbc}
\title{Create a connection to databricks}
\usage{
databricks_jdbc(adress, port, organization, cluster, token)
}
\arguments{
\item{adress}{URL to the cluster}

\item{port}{port}

\item{organization}{organization id}

\item{cluster}{cluster ID}

\item{token}{access token}
}
\description{
Create a connection to databricks
}
